% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.nonstat.R
\name{is.nonstat}
\alias{is.nonstat}
\title{Test for nonstationarity}
\usage{
is.nonstat(tseries, nEp = 2, cut.psr1 = 1.1, cut.psr2 = 1.01, span = 3)
}
\arguments{
\item{tseries}{a numerical vector}

\item{nEp}{number of epochs (in which time series is cut for PSR calculation)}

\item{cut.psr1}{threshold for PSR value that indicates mean non-stationarity}

\item{cut.psr2}{threshold for PSR value that indicates variance non-stationarity}

\item{span}{numerical value that is passed to the \code{loess} function}
}
\value{
a logical scalar indicating whether the time series is stationary (\code{TRUE}) or not (\code{FALSE})
}
\description{
This function tests whether a time series is stationary or not
}
\examples{
set.seed( 8332278 )
x <- rnorm( 50 )
is.nonstat( x )
}
\references{
Zitzmann, S., Lindner, C., Lohmann, J. F., & Hecht, M. (2024). A novel nonvisual procedure for screening for nonstationarity in time series as obtained from intensive longitudinal designs [Preprint]. https://www.researchgate.net/publication/384354932_A_Novel_Nonvisual_Procedure_for_ Screening_for_Nonstationarity_in_Time_Series_as_Obtained_from_Intensive_Longitudinal_Designs
}
