% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.nonstat.R
\name{is.nonstat}
\alias{is.nonstat}
\title{Test for nonstationarity}
\usage{
is.nonstat(tseries, nEp = 2, cut.psr1 = 1.1, cut.psr2 = 1.01, span = 3)
}
\arguments{
\item{tseries}{a numerical vector}

\item{nEp}{number of epochs (in which time series is cut for PSR calculation)}

\item{cut.psr1}{threshold for PSR value that indicates mean non-stationarity}

\item{cut.psr2}{threshold for PSR value that indicates variance non-stationarity}

\item{span}{numerical value that is passed to the \code{loess} function}
}
\value{
a logical scalar indicating whether the time series is stationary (\code{TRUE}) or not (\code{FALSE})
}
\description{
This function tests whether a time series is stationary or not
}
\examples{
set.seed( 1234 )
x <- rnorm( 100 )
is.nonstat( x )
}
